クラスをlib/autoloads/import_csv.rbに移動
