+++ db/seeds.rb
@@ -15,32 +15,3 @@ User.find_or_create_by!(email: EMAIL) do |user|
   user.password = PASSWORD
   puts "ユーザーの初期データインポートに成功しました。"
 end
-
-
-require "csv"
-class ImportCsv
-
-  # CSVデータのパスを引数として受け取り、インポート処理を実行
-  def self.import(path)
-    # インポートするデータを格納するための空配列
-    list = []
-
-    # CSVファイルからインポートしたデータを格納
-    CSV.foreach(path, headers: true) do |row|
-      list << row.to_h
-    end
-    # メソッドの戻り値をインポートしたデータの配列とする
-    list
-  end
-
-  def self.text_data
-    # importメソッドを呼び出し、テキストデータの配列を生成
-    list = import('db/csv_data/text_data.csv')
-    Text.destroy_all #データの初期化
-
-    puts "インポート処理を開始"
-    Text.create!(list)
-    puts "インポート完了!!"
-  end
-
-end
